refactor(ioc): rename syntax-transformer to ts-syntax-transformer
