+++ ioc/analysis/ts-syntax-transformer.ts
