+++ test/unit/ioc/code-generator.spec.ts
@@ -13,7 +13,7 @@ import * as ts from 'typescript';
 import { raw } from '../../fixture/raw';
 import * as AST from '../../../ioc/analysis/ast';
 import { StaticModuleConfiguration } from '../../../ioc/static-module-configuration';
-import { TypeScriptSyntaxTransformer } from '../../../ioc/analysis/syntax-transformer';
+import { TypeScriptSyntaxTransformer } from '../../../ioc/analysis/ts-syntax-transformer';
 import { InjectorBuilder } from '../../../ioc/injector';
 import { Node } from '../../../ioc/graph';
 import { IObjectContext } from '../../../ioc/composition/interfaces';
