+++ test/unit/ioc/injector-builder.spec.ts
@@ -43,7 +43,7 @@ import { IContext } from '../../../ioc/interfaces';
 import * as ts from 'typescript';
 import { raw } from '../../fixture/raw';
 import * as AST from '../../../ioc/analysis/ast';
-import { TypeScriptSyntaxTransformer } from '../../../ioc/analysis/syntax-transformer';
+import { TypeScriptSyntaxTransformer } from '../../../ioc/analysis/ts-syntax-transformer';
 
 describe('InjectorBuilder', () => {
   let sut: InjectorBuilder;
